+++ @xen-orchestra/lite/src/components/infra/InfraPoolList.vue
@@ -18,7 +18,6 @@
               <UiButtonIcon
                 accent="brand"
                 icon="action:more-actions"
-                variant="tertiary"
                 size="small"
                 :selected="isOpen"
                 @click="open($event)"