+++ @xen-orchestra/lite/src/modules/host/components/actions/download/HostDownloadButton.vue
@@ -1,18 +1,35 @@
 <template>
-  <MenuItem icon="action:download" @click="download()">{{ t('action:download-bugtools-archive') }}</MenuItem>
+  <MenuItem
+    icon="action:download"
+    :busy="isFetching"
+    :disabled="(isReady && poolMaster === undefined) || hasError"
+    @click="download()"
+  >
+    {{ t('action:download-bugtools-archive') }}
+  </MenuItem>
 </template>
 
 <script lang="ts" setup>
+import type { XenApiHost } from '@/libs/xen-api/xen-api.types.ts'
+import { useHostStore } from '@/stores/xen-api/host.store'
 import MenuItem from '@core/components/menu/MenuItem.vue'
-import { downloadFile } from '@core/utils/download-file.utils.ts'
-import type { XenApiHost } from '@vates/types'
+import { downloadBugTools } from '@core/utils/download-bugtools.utils'
+import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
-const { hostIp } = defineProps<{ hostIp: XenApiHost['address'] }>()
+const { hostOpaqueRef } = defineProps<{ hostOpaqueRef: XenApiHost['$ref'] }>()
 
 const { t } = useI18n()
 
+const { getByOpaqueRef, isReady, isFetching, hasError } = useHostStore().subscribe()
+
+const poolMaster = computed(() => getByOpaqueRef(hostOpaqueRef))
+
 const download = () => {
-  downloadFile(`http://${hostIp}/system-status?output=tar.bz2`, 'system-status.tar.bz2')
+  if (poolMaster.value?.address === undefined) {
+    return
+  }
+
+  downloadBugTools(poolMaster.value.address)
 }
 </script>