+++ @xen-orchestra/lite/src/modules/pool/components/actions/PoolTreeActions.vue
@@ -1,10 +1,10 @@
 <template>
-  <PoolMoreActions :pool />
+  <PoolDownloadButton :master-host="pool.master" />
 </template>
 
 <script setup lang="ts">
 import type { XenApiPool } from '@/libs/xen-api/xen-api.types.ts'
-import PoolMoreActions from '@/modules/pool/components/actions/PoolMoreActions.vue'
+import PoolDownloadButton from '@/modules/pool/components/actions/download/PoolDownloadButton.vue'
 
 defineProps<{
   pool: XenApiPool