+++ @xen-orchestra/web-core/lib/utils/download-file.utils.ts
@@ -1,7 +1,12 @@
-export function downloadFile(url: string, downloadName: string) {
+export function downloadFile(url: string, downloadName: string, openInNewTab: boolean = true) {
   const link = document.createElement('a')
   link.href = url
   link.download = downloadName
+
+  if (openInNewTab) {
+    link.target = '_blank'
+  }
+
   document.body.appendChild(link)
   link.click()
   document.body.removeChild(link)