+++ @xen-orchestra/web/src/modules/host/components/actions/download/HostDownloadButton.vue
@@ -1,18 +1,33 @@
 <template>
-  <MenuItem icon="action:download" @click="download()">{{ t('action:download-bugtools-archive') }}</MenuItem>
+  <MenuItem
+    icon="action:download"
+    :busy="areHostsFetching"
+    :disabled="(areHostsReady && host === undefined) || hasHostFetchError"
+    @click="download()"
+  >
+    {{ t('action:download-bugtools-archive') }}
+  </MenuItem>
 </template>
 
 <script lang="ts" setup>
-import type { FrontXoHost } from '@/modules/host/remote-resources/use-xo-host-collection.ts'
+import { useXoHostCollection, type FrontXoHost } from '@/modules/host/remote-resources/use-xo-host-collection.ts'
 import MenuItem from '@core/components/menu/MenuItem.vue'
-import { downloadFile } from '@core/utils/download-file.utils.ts'
+import { downloadBugTools } from '@core/utils/download-bugtools.utils'
 import { useI18n } from 'vue-i18n'
 
-const { hostIp } = defineProps<{ hostIp: FrontXoHost['address'] }>()
+const { hostId } = defineProps<{ hostId: FrontXoHost['id'] }>()
 
 const { t } = useI18n()
 
+const { getHostById, areHostsFetching, areHostsReady, hasHostFetchError } = useXoHostCollection()
+
+const host = getHostById(hostId)
+
 const download = () => {
-  downloadFile(`http://${hostIp}/system-status?output=tar.bz2`, 'system-status.tar.bz2')
+  if (!host) {
+    return
+  }
+
+  downloadBugTools(host.address)
 }
 </script>