+++ @xen-orchestra/web/src/modules/treeview/components/PoolTreeItem.vue
@@ -15,7 +15,6 @@
             <UiButtonIcon
               accent="brand"
               icon="action:more-actions"
-              variant="tertiary"
               size="small"
               :selected="isOpen"
               @click="open($event)"